refactor: move commands into package
