+++ cmd/cli/commands/version.go
@@ -1,0 +1,14 @@
+package commands
+
+import (
+	"fmt"
+	"github.com/spf13/cobra"
+)
+
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Show version",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Printf("Version: %s\r\n", version)
+	},
+}
